refactor(gear): extract GearPicker core to a neutral src/components/gear (B1)
Section B1 of the build⇄loadout bridge plan: lift the set-aware GearPicker
core out of build-editor into a neutral location so non-build-editor surfaces
(the Loadout Manager, in B2) can reuse it without creating a
loadout-manager → build-editor component import cycle.

- Move GearPickerDialog → src/components/gear/GearPickerDialog.tsx and export
  its props type. The component is already fully presentational (keyed on item
  ids via targetSlot/currentItemId + onSelect(itemId), no Redux), so its only
  build-editor deps are leaf data/hook modules (gearSetRegistry, setArmorWeights,
  useDebouncedValue) — reaching those creates no cycle (the plan's constraint is
  specifically: loadout-manager must not import build-editor *components*).
- Add a src/components/gear barrel.
- Leave a thin deprecated re-export shim at the old pickers/GearPicker.tsx path
  for back-compat; point EquipmentPicker at the neutral path.
- Move the 5 GearPicker test suites alongside the core so they exercise it
  directly in its new home (icon-race / stale-icon / mixed-weapons /
  offhand-dual-wield / icon-error coverage all pass unchanged).

No behavior change — pure relocation. Full tsc, ESLint, and the moved picker
suites pass.
