+++ src/components/gear/__tests__/GearPicker.staleIcon.test.tsx
@@ -28,7 +28,7 @@ jest.mock('@features/loadout-manager/utils/itemIconResolver', () => {
   };
 });
 
-import { GearPickerDialog } from '../GearPicker';
+import { GearPickerDialog } from '../GearPickerDialog';
 
 const renderPicker = (slot: 'weapon' | 'chest') =>
   render(
