+++ src/components/gear/index.ts
@@ -1,0 +1,2 @@
+export { GearPickerDialog } from './GearPickerDialog';
+export type { GearPickerDialogProps } from './GearPickerDialog';
